next

+++ RC_FFMPEG/app/src/main/java/cc/dewdrop/ffplayer/Commander.java
@@ -88,7 +88,7 @@ public class Commander {
     static byte fpv_addr;
     static short fpv_port;
     static short fpv_code=0;
-    static byte fpv_zoom=0;
+    static int fpv_zoom=0;
 
     //static boolean exit_main=false;
 
@@ -250,7 +250,7 @@ public class Commander {
             buf[i++]=fpv_addr;
             buf[i++]=(byte)(0xff&fpv_port);
             buf[i++]=(byte)(0xff&(fpv_port>>8));
-            buf[i++]=fpv_zoom;
+            buf[i++]=(byte)fpv_zoom;
             buf[i++]=(byte)(0xff&fpv_code);
             buf[i++]=(byte)(0xff&(fpv_code>>8));
             fpv_code=0;
